More fixes from updating aws-crt-python (#6)
* Allow clang to use a default version

* Fixed logic for when project is reloaded after download

* Added better debugging of project resolution

* Added robust clang/gcc compiler detection

* Added a sanity test before kicking off docker jobs

* Added support for apple clang

* Fixed spec -> toolchain in cmake

* Use vswhere to find VS on windows

* Load scripts at any level in .builder

* Exposed more classes via Builder

* Fixed buffering/output issues with exec()

* Make sure submodules are always synced

* Dump inspect output always

* Fail job if a script command has a non-0 exit code

* Apply configs from platform, then host

* Fixed up arg parsing for actions, added --compiler and --target

* Fixed default compiler output when compiler is not installed yet

* Export compiler after installing it

* moved compiler concerns into InstallCompiler

* Fixed script output when no new classes are loaded

+++ builder/env.py
@@ -11,34 +11,19 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
-import glob
 import os
 import subprocess
 import sys
 
 from actions.git import DownloadSource
 from project import Project
-from scripts import Scripts
 from shell import Shell
-
-
-def looks_like_code(path):
-    git_dir = os.path.isdir(os.path.join(path, '.git'))
-    if git_dir:
-        return True
-    # maybe not committed yet?
-    readme = glob.glob(os.path.join(path, 'README.*'))
-    if readme:
-        return True
-    return False
 
 
 class Env(object):
     """ Encapsulates the environment in which the build is running """
 
     def __init__(self, config={}):
-        self._projects = {}
-
         # DEFAULTS
         self.dryrun = False  # overwritten by config
         # default the branch to whatever the current dir+git says it is
@@ -53,41 +38,50 @@ class Env(object):
             self.shell = Shell(self.dryrun)
 
         # build environment set up
-        self.source_dir = os.environ.get(
-            "CODEBUILD_SRC_DIR", self.shell.cwd())
-        self.build_dir = os.path.join(self.source_dir, self.args.build_dir)
+        self.source_dir = os.path.abspath(self.args.build_dir)
+        self.build_dir = os.path.join(self.source_dir, 'build')
         self.deps_dir = os.path.join(self.build_dir, 'deps')
         self.install_dir = os.path.join(self.build_dir, 'install')
-        self.launch_dir = self.shell.cwd()
+        self.launch_dir = os.path.abspath(self.shell.cwd())
+
+        Project.search_dirs = [
+            self.launch_dir,
+            self.build_dir,
+            self.source_dir,
+            self.deps_dir,
+        ]
 
         print('Source directory: {}'.format(self.source_dir))
 
-        # Once initialized, switch to the build dir for everything else
-        self.shell.rm(self.build_dir)
+        if os.path.exists(self.build_dir):
+            self.shell.rm(self.build_dir)
         self.shell.mkdir(self.build_dir)
-        self.shell.cd(self.build_dir)
 
         # default the project to whatever can be found, or convert
         # project from a name to a Project
         if not hasattr(self, 'project'):
-            self.project = self._default_project()
+            self.project = Project.default_project()
 
         if not self.project and not self.args.project:
-            print('No project specified and no project found in current directory')
-            sys.exit(1)
-        project_name = self.project if self.project else self.args.project1
+            return
+
+        project_name = self.project if self.project else self.args.project
         # Ensure that the specified project exists, this may return a ref or the project if
         # it is present on disk
-        project = self.find_project(project_name)
+        project = Project.find_project(project_name)
         if not project.path:  # got a ref
             print('Project {} could not be found locally, downloading'.format(
                 project.name))
             DownloadSource(
-                project=project, branch=self.branch).run(self)
+                project=project, branch=self.branch, path=self.build_dir).run(self)
             # Now that the project is downloaded, look it up again
-            project = self.find_project(project.name)
+            project = Project.find_project(project.name)
             assert project.path
         self.project = project
+
+        # Once initialized, switch to the source dir before running actions
+        if self.project:
+            self.shell.cd(self.project.path)
 
     @staticmethod
     def _get_git_branch():
@@ -122,107 +116,3 @@ class Env(object):
             return branch
 
         return None
-
-    def _cache_project(self, project):
-        self._projects[project.name] = project
-        Scripts.load(project.path)
-        return project
-
-    def _default_project(self):
-        project = self._project_from_cwd()
-        if project:
-            return self._cache_project(project)
-        if not self.args.project:
-            print(
-                "Multiple projects available and no project (-p|--project) specified")
-            print("Available projects:", ', '.join(
-                [p.__name__ for p in Project.__subclasses__()]))
-            sys.exit(1)
-
-        project_name = self.args.project
-        projects = Project.__subclasses__()
-        for project_cls in projects:
-            if project_cls.__name__ == project_name:
-                project = project_cls()
-                project.path = self.shell.cwd()
-                return self._cache_project(project)
-        print("Could not find project named {}".format(project_name))
-        sys.exit(1)
-
-    def _project_from_cwd(self, name_hint=None):
-        project_config = None
-        project_config_file = os.path.abspath("builder.json")
-        if os.path.exists(project_config_file):
-            import json
-            with open(project_config_file, 'r') as config_fp:
-                try:
-                    project_config = json.load(config_fp)
-                except Exception as e:
-                    print("Failed to parse config file",
-                          project_config_file, e)
-                    sys.exit(1)
-                return self._cache_project(Project(**project_config, path=self.shell.cwd()))
-
-        # load any builder scripts and check them
-        print("No project file found at {}, loading scripts".format(
-            project_config_file))
-        Scripts.load()
-        projects = Project.__subclasses__()
-        project_cls = None
-        if len(projects) == 1:
-            project_cls = projects[0]
-        elif name_hint:  # if there are multiple projects, try to use the hint if there is one
-            for p in projects:
-                if p.__name__ == name_hint:
-                    project_cls = p
-
-        if project_cls:
-            project = project_cls()
-            project.path = self.shell.cwd()
-            return self._cache_project(project)
-
-        return None
-
-    def find_project(self, name):
-        """ Finds a project, either on disk, or makes a virtual one to allow for acquisition """
-        project = self._projects.get(name, None)
-        if project:
-            return project
-
-        sh = self.shell
-        search_dirs = (os.path.abspath(self.launch_dir), os.path.abspath(
-            os.path.join('.', name)), self.source_dir, os.path.join(self.deps_dir, name))
-
-        for search_dir in search_dirs:
-            if (os.path.basename(search_dir) == name) and os.path.isdir(search_dir):
-                sh.pushd(search_dir)
-                project = self._project_from_cwd(name)
-                sh.popd()
-
-                # might be a project without a config
-                if not project and looks_like_code(search_dir):
-                    project = self._cache_project(
-                        Project(name=name, path=search_dir))
-                    return project
-
-        # Enough of a project to get started, note that this is not cached
-        return Project(name=name)
-
-    def _find_compiler_tool(self, name, versions):
-        for version in versions:
-            for pattern in ('{name}-{version}', '{name}-{version}.0'):
-                exe = pattern.format(name=name, version=version)
-                path = self.shell.where(exe)
-                if path:
-                    return path, version
-        return None, None
-
-    def find_gcc_tool(self, name, version=None):
-        """ Finds gcc, gcc-ld, gcc-ranlib, etc at a specific version, or the latest one available """
-        versions = [version] if version else list(range(8, 5, -1))
-        return self._find_compiler_tool(name, versions)
-
-    def find_llvm_tool(self, name, version=None):
-        """ Finds clang, clang-tidy, lld, etc at a specific version, or the latest one available """
-        versions = [version] if version else list(range(10, 6, -1))
-        return self._find_compiler_tool(name, versions)
